+++ packages/providers/lib/websocket-provider.js
@@ -331,8 +331,9 @@ var WebSocketProvider = /** @class */ (function (_super) {
                         _a.sent();
                         _a.label = 2;
                     case 2:
-                        // Hangup (navigating away from the page that opened the connection)
-                        this._websocket.close(1001);
+                        // Hangup
+                        // See: https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent#Status_codes
+                        this._websocket.close(1000);
                         return [2 /*return*/];
                 }
             });